Create Order List Api

+++ database/migrations/2024_07_25_183943_create_order_items_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->foreign( 'orderId' )->references( 'id' )->on( 'orders' );
             $table->unsignedBigInteger( 'productId' );
             $table->foreign( 'productId' )->references( 'id' )->on( 'products' );
-            $table->integer( 'quantity' );
-            $table->integer( 'unitPrice' );
-            $table->integer( 'total' );
+            $table->float( 'quantity' );
+            $table->float( 'unitPrice' );
+            $table->float( 'total' );
             $table->timestamps();
         } );
     }
